+++ src/pages/Hero/Hero.jsx
@@ -1,12 +1,12 @@
 import styles from './Hero.module.css'
-import HeroHeader from "components/HeroHeader/HeroHeader";
+import GuestHeader from "components/GuestHeader/GuestHeader";
 import HeroContainer from 'components/HeroContainer/HeroContainer';
 import HeroDescription from 'components/HeroDescription/HeroDescription';
 
 const Hero = () => {
     return (
         <div className={styles.heroContainer}>
-            <HeroHeader/>
+            <GuestHeader/>
             <HeroContainer/>
             <HeroDescription/>
         </div>
